AMB-809 create and consume  javascript library
